Fix remaining compilation errors
- Fix PCCameraController: use existing field names (worldMapPanSpeed, territoryOrbitSpeed)
- Add _apexCoins field to PCResourceSystem
- Fix UIUINotificationType -> UINotificationType in PCTerritoryBridge
- Fix UIUnityEngine.UI.Outline -> UnityEngine.UI.Outline (malformed replacement)
- Fix all Outline variable declarations to use UnityEngine.UI.Outline
- Fix BarracksPanel: cast GetResource float to int
- Fix ShowResourceGained: use string 'Gold' instead of ResourceType.Gold
- Fix TutorialSystem HighlightPulse class Outline references

+++ unity/ApexCitadels/Assets/Scripts/PC/UI/BarracksPanel.cs
@@ -553,8 +553,8 @@ namespace ApexCitadels.PC.UI
                 // Check resources
                 if (PCResourceSystem.Instance != null)
                 {
-                    int gold = PCResourceSystem.Instance.GetResource(ResourceType.Gold);
-                    int stone = PCResourceSystem.Instance.GetResource(ResourceType.Stone);
+                    int gold = (int)PCResourceSystem.Instance.GetResource(ResourceType.Gold);
+                    int stone = (int)PCResourceSystem.Instance.GetResource(ResourceType.Stone);
                     
                     if (gold < def.GoldCost || stone < def.StoneCost)
                     {
